+++ src/RunningAverage.h
@@ -5,13 +5,13 @@ class RunningAverage {
 	public:
 		RunningAverage(unsigned int size);
 		~RunningAverage();
-		void fill(int value);
-		float append(int value);
+		void fill(float value);
+		float append(float value);
 	private:
 		unsigned int mSize;
 		unsigned int mI;
-		int mTotal;
-		int *mValues;
+		float mTotal;
+		float *mValues;
 };
 
 #endif
